Add Codex submodule to lib/codex and embed Codex demos in Sandbox
Codex is developed in a stand-alone repository, but (for now at
least) most components developed there are destined to be used
inside of MediaWiki. We need a good way to test out Codex components
(including experimental unreleased ones) inside a MW environment
to catch bugs and ensure correct appearance/behavior.

To do this, the entire Codex repo has been added here as a submodule
under lib/codex. As needed, developers can update the submodule
to the latest upstream commit and run a build script to bundle up
the latest set of component demos. Those demos can then be copied
and used in the VueTest sandbox page. This is intended to support
a manual QA testing process every time a new Codex component is
developed.

To update the bundled codex-demos to the latest version available,
developers should do two things:

1. Update the Codex submodule: git submodule update --remote --merge
2. Run the build script: npm run codex:build-demos

This will install all submodule dependencies, run the submodule
build commands, and copy the resulting files into the resources/
folder. Updates to the bundled JS and CSS files should be committed
to source so that they are available in places like patchdemo.

* Adds Codex as a submodule under lib/codex
* Replaces the Codex component demos with bundles from
  Codex (local build script needs to be run to generate)
* Introduces some new build scripts in package.json to simplify
  the process of updating demos.

Bug: T318842

Depends-On: I294f01f766a767d373a8e4a1155cc91c06123ee5
Change-Id: I671e92bdc484a786c474418b8a6caa2fc83c6627