Update Codex to v1.2.1
Also replaces the Codex demo code with code generated from
the latest version. This required one minor change to the
build scripts because the demos (which are built inside of
the codex-docs package within Codex) now default to ESM
output and we need to ask for the .cjs file explicilty
for use inside of MediaWiki.

Change-Id: Id5849ef95f101e825e0feee8f86fe7ea2541506b